remove unnecessary downcasting of int

+++ src/main/java/core/basesyntax/FindNumbers.java
@@ -14,7 +14,7 @@ public class FindNumbers {
         String[] tableString = (text.replaceAll("[^-?\\d+]", " ").trim().split(" +"));
         int[] tableInt = new int[tableString.length];
         for (int i = 0; i < tableInt.length; i++) {
-            tableInt[i] = (int) Integer.parseInt(tableString[i]) * 2;
+            tableInt[i] = Integer.parseInt(tableString[i]) * 2;
         }
         return tableInt;
     }
